feat: 一些功能细节优化

+++ biomass-backend/src/main/java/com/xmu/biomass/common/entity/BaseEntity.java
@@ -33,5 +33,5 @@ public class BaseEntity {
     /**
      * 删除标志
      */
-    private Double deleteAt;
+    private Long deleteAt;
 }
