+++ biomass-backend/src/main/java/com/xmu/biomass/common/vo/BaseVo.java
@@ -30,5 +30,5 @@ public class BaseVo {
     /**
      * 删除标志
      */
-    private Double deleteAt;
+    private Long deleteAt;
 }
